aba calls charout for both processes && formating done

+++ 01_processes/src/aba.cpp
@@ -14,32 +14,66 @@ using namespace std;
 
 int main() {
 
-    chrono::milliseconds sleeptime(500);
+    
 
-    int i{0};
+    chrono::milliseconds sleeptime(3000);
+
+    cout << "waiting for 3 seconds" << endl;
 
     auto pid{fork()};
 
     if (pid == 0) {
-        execl("./charout", "charout", "A", nullptr);
-        if (errno){
-            cerr << strerror(errno) << endl;
-            quick_exit(1);
+        const char* A{getenv("ABA_LETTER_A")};
+        if (A){
+            execl("./charout", "charout", A, nullptr);
+            if (errno){
+                cerr << "starting charout failed: "<< strerror(errno) << endl;
+                quick_exit(1);
+            }
+        }
+        else{
+            cout << "Letter A not set" << endl;
         }
         
+        
+        
     } else {
-        int status{};
-        while (i++ < 6){
-            cout << "B" << flush;
+
+        auto processpid{fork()};
+
+        if (processpid == 0) {
+            const char* B{getenv("ABA_LETTER_B")};
+            if (B){
+                execl("./charout", "charout", "B", nullptr);
+                if (errno){
+                cerr << "starting charout failed: " << strerror(errno) << endl;
+                quick_exit(1);
+                }
+            } 
+            else {
+                cout << "Letter B not set" << endl;
+            }
+        }
+        else {
+            int status{};
+            int status2{};
+            
             this_thread::sleep_for(sleeptime);
+
+            cout << "\nkilling both subprocesses with pids " << pid << " and " << processpid << endl;
+            kill(pid, SIGKILL);
+            kill(processpid, SIGKILL);
+
+            cout << "for both subprocesses to be dead" << endl;
+            waitpid(pid, &status, 0);
+            waitpid(processpid, &status2, 0);
+
+            cout << "subprocess " << pid << " exited with " << WEXITSTATUS(status) << endl;
+            cout << "subprocess " << processpid << " exited with" << WEXITSTATUS(status2) << endl;
+            
         }
 
-        kill(pid, SIGKILL);
-        waitpid(pid, &status, 0);
-
-
-
-        cout << "subprocess " << pid << " exited with " << WEXITSTATUS(status) << endl;
+        
         
     }
